### 2016-11-14 #112 @whedon
[PRE REVIEW]: Multilocus sequence typing by blast from de novo assemblies against PubMLST
**Submitting author:** @andrewjpage (<a href="http://orcid.org/0000-0001-6919-6062">Andrew Page</a>)
**Repository:** <a href="https://github.com/sanger-pathogens/mlst_check" target ="_blank">https://github.com/sanger-pathogens/mlst_check</a>
**Version:** v2.1.1630910
