### 2016-11-14 #112 @pjotrp
@happykhan: feel free to pitch in with review
