### 2016-11-18 #118 @happykhan
I can confirm that the program works as described, on real world data and meets all of the requirements stated here.
