### 2016-11-22 #118 @andrewjpage
Thanks @happykhan
If you need some alternative suggestions for reviewers, heres some researchers who have written MLST code & are active on GitHub (and I have never published with):

@lskatz (CDC - Centers for Disease Control) https://github.com/lskatz/lyve-MLST
@aunderwo (PHE - Public Health England) https://github.com/phe-bioinformatics/MOST
@jacarrico (Universidade de Lisboa, Portugal) https://github.com/jacarrico/MLSTtrimm & ReMatCh
