### 2016-11-26 #118 @harmn
I reviewed mlst_check: it installs and runs. Some issues/comments below, in general the documentation is a bit limited.

**Installation instructions**
Installation works fine on Ubuntu 16.04.1 LTS, but it does require root access/sudo rights or a docker installation. Might be good to mention that.

**Example usage**
I could not find the sample files described in the README.md file for the docker container (/data/sample1.fa etc.). I used the Salmonella files in "mlst_check/example/input_data" to test the software.

**Documentation**
The used methods are poorly described (as far as I can tell), for instance I had to look in the docstrings to find out what is used to find the best hit: "The best hit has the greatest number of matching bases.  If two hits have the same number of matching bases, the one with the greater percentage identity is selected."

I had to search the code for the meaning of the tilde in the ST column, "Add tilde to matches which are not 100%". Would be good to add that to the README.md.

What is exactly expected in the input FASTA files (I assume complete genomes)?
What is the output if your sequences have no hits with the selected scheme?

**Automated tests**
There are  quite a number test scripts, but I could not find any description of the tests. I ran all and found no problems.

**References**
Page et al. does not have a doi (should be http://dx.doi.org/10.1099/mgen.0.000083)
