### 2016-11-27 #118 @pjotrp
@andrewjpage Do you mind addressing mentioned issues?
