### 2016-11-28 #118 @andrewjpage
@harmn Thanks for the comprehensive review.
@pjotrp Yes I will address all of the issues, but it will take me a few days.
