### 2016-11-29 c94b4cb
updates from review
