### 2016-11-29 b88ae75
fix link to example files
