### 2016-11-30 af1d53d
more descriptive tests
