### 2016-11-30 #118 @andrewjpage
Hi @harmn
Thanks for reviewing this submission.

**Installation instructions
Installation works fine on Ubuntu 16.04.1 LTS, but it does require root access/sudo rights or a docker installation. Might be good to mention that.**

   I have clarified that root access is needed, and also added instructions for installing with HomeBrew/LinuxBrew (which doesn't need root).

**Example usage
I could not find the sample files described in the README.md file for the docker container (/data/sample1.fa etc.). I used the Salmonella files in "mlst_check/example/input_data" to test the software.**

   I have added test files to the docker container so that the example command works out of the box.

**Documentation
The used methods are poorly described (as far as I can tell), for instance I had to look in the docstrings to find out what is used to find the best hit: "The best hit has the greatest number of matching bases. If two hits have the same number of matching bases, the one with the greater percentage identity is selected."**

**I had to search the code for the meaning of the tilde in the ST column, "Add tilde to matches which are not 100%". Would be good to add that to the README.md.**

**What is exactly expected in the input FASTA files (I assume complete genomes)?
What is the output if your sequences have no hits with the selected scheme?**

   I have extended README documentation to clarify the methods used, the meaning of the various notations and extended the examples.

**Automated tests
There are quite a number test scripts, but I could not find any description of the tests. I ran all and found no problems.**

   By default the tests produce minimal output when running, but can be run in a more verbose mode 'dzil test --test-verbose' which gives a text string for each test. I have improved the messages coming out, and added more descriptive text to blocks of tests to make it clearer whats going on. Also I have set TravisCI to run the tests in verbose mode so that its easier to see.

**References
Page et al. does not have a doi (should be http://dx.doi.org/10.1099/mgen.0.000083)**

   Thanks this has been corrected.
