### 2016-12-01 #118 @harmn
@andrewjpage nice work
@pjotrp happy to accept this
