### 2016-12-05 #118 @andrewjpage
I've added it to Figshare as: https://dx.doi.org/10.6084/m9.figshare.4285097.v1
