### 2016-12-15 #118 @arfon
Thanks for the heads up @genomematt - this should be fixed now.
